Add implementation of two-adic u256 prime field
