+++ math/src/field/fields/mod.rs
@@ -1,5 +1,5 @@
 /// Implementation of two-adic prime fields to use with the Fast Fourier Transform (FFT).
-pub mod friendly_fft;
+pub mod fft_friendly;
 pub mod montgomery_backed_prime_fields;
 /// Implementation of the Goldilocks Prime field (p = 2^448 - 2^224 - 1)
 pub mod p448_goldilocks_prime_field;
